collection-service/functions/Search/requirements.txt
python-jose
boto3
aws_xray_sdk